db connection middleware issues
