+++ package.json
@@ -2,12 +2,12 @@
   "name": "rmbrme-server",
   "version": "1.0.0",
   "description": "App to help people remember things about their friends and family",
-  "main": "dist/server.js",
+  "main": "src/server.ts",
   "scripts": {
     "tsc": "tsc",
     "predev": "tsc",
     "test": "mocha --require test/setup.js",
-    "dev": "nodemon src/server.ts",
+    "dev": "nodemon -r ts-node/register src/server.ts",
     "migrate": "postgrator --config postgrator-config.js",
     "migrate:production": "env SSL=true DATABASE_URL=$(heroku config:get DATABASE_URL) npm run migrate",
     "migrate:test": "env NODE_ENV=test npm run migrate",
