+++ src/server.ts
@@ -13,6 +13,7 @@ import rmbrRouter from "./rmbr/rmbr-router";
 import authRouter from "./auth/auth-router";
 import userRouter from "./user/user-router";
 import { addDbMiddleware } from "./middleware/add-db";
+import {db} from "./database/connect";
 
 // main express root
 const app = express();
